modify entities and prepopulation function

+++ app/src/main/java/inas/anisha/skripsi_app/data/db/entity/CycleEntity.kt
@@ -11,6 +11,6 @@ import kotlinx.android.parcel.Parcelize
 data class CycleEntity(
     @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Long,
     @ColumnInfo(name = "number") val number: Int = 0,
-    @ColumnInfo(name = "completion") val completion: Long = 0,
+    @ColumnInfo(name = "completion") val completion: Int = 0,
     @ColumnInfo(name = "reflection") val reflection: String = ""
 ) : Parcelable
